Terminado MVC

+++ spring-boot-backend-apirest/src/main/java/springboot/backend/apirest/models/dao/IOrderDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import springboot.backend.apirest.models.entity.Order1;
 
-public interface IOrderDao extends PagingAndSortingRepository<Order1, Long>{
+public interface IOrderDao extends PagingAndSortingRepository<Order1, Integer>{
 
 }
